Server rooms
Rooms provide area of interest management for clients, allowing clients to listen only on specific object updates through rooms. For now, the rooms are represented through a single UInt32 identifier, and clients "join" and "leave" rooms by requesting it to the server.

Clients can connect to multiple rooms simultaneously, up to 10 rooms (hard-coded right now). Every networked object is part of one room each, and can be transferred to other rooms. When transferred, clients who hasn't seen the object previously can initialize it, and clients who does not listen to the room it was transferred to can remove it.

A more complex room system will be required when the server needs to do additional load balancing.